+++ sport-service/src/main/java/br/edu/atitus/sportservice/controllers/SportController.java
@@ -3,6 +3,7 @@ package br.edu.atitus.sportservice.controllers;
 import br.edu.atitus.sportservice.DTOs.SportDTO;
 import br.edu.atitus.sportservice.entities.SportEntity;
 import br.edu.atitus.sportservice.services.SportService;
+import org.apache.tomcat.websocket.AuthenticationException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,7 +12,7 @@ import java.util.List;
 import java.util.UUID;
 
 @RestController
-@RequestMapping("/sports")
+@RequestMapping("/ws/sports")
 public class SportController {
 
     private final SportService sportService;
@@ -20,13 +21,8 @@ public class SportController {
         this.sportService = sportService;
     }
 
-    @PostMapping("/create")
-    public ResponseEntity<SportEntity> createSport(@RequestBody SportDTO sportDTO) throws Exception {
-        SportEntity sport = new SportEntity();
-        BeanUtils.copyProperties(sportDTO, sport);
-
-        var created = sportService.createSport(sport);
-        return ResponseEntity.ok(created);
+    private boolean validadeUserType(Long userType) {
+        return userType != 0;
     }
 
     @GetMapping("/{sportId}")
@@ -45,12 +41,44 @@ public class SportController {
     }
 
     @DeleteMapping("/delete/{sportId}")
-    public void deleteSport(@PathVariable UUID sportId) {
+    public void deleteSport(
+            @PathVariable UUID sportId,
+            @RequestHeader("X-User-Id") Long userId,
+            @RequestHeader("X-User-Email") String userEmail,
+            @RequestHeader("X-User-Type") Long userType
+    ) throws Exception
+    {
+        if (validadeUserType(userType)) throw new AuthenticationException("Usuário sem nível de acesso");
         sportService.deleteSport(sportId);
     }
 
+    @PostMapping("/create")
+    public ResponseEntity<SportEntity> createSport(
+            @RequestBody SportDTO sportDTO,
+            @RequestHeader("X-User-Id") Long userId,
+            @RequestHeader("X-User-Email") String userEmail,
+            @RequestHeader("X-User-Type") Long userType
+    ) throws Exception
+    {
+        if (validadeUserType(userType)) throw new AuthenticationException("Usuário sem nível de acesso");
+
+        SportEntity sport = new SportEntity();
+        BeanUtils.copyProperties(sportDTO, sport);
+
+        var created = sportService.createSport(sport);
+        return ResponseEntity.ok(created);
+    }
+
     @PutMapping("/update/{sportId}")
-    public SportEntity updateSport(@PathVariable UUID sportId, @RequestBody SportDTO sportDTO) throws Exception {
+    public SportEntity updateSport(
+            @PathVariable UUID sportId,
+            @RequestBody SportDTO sportDTO,
+            @RequestHeader("X-User-Id") Long userId,
+            @RequestHeader("X-User-Email") String userEmail,
+            @RequestHeader("X-User-Type") Long userType
+    ) throws Exception
+    {
+        if (validadeUserType(userType)) throw new AuthenticationException("Usuário sem nível de acesso");
         return sportService.updateSport(sportId, sportDTO);
     }
 }
