+++ team-service/src/main/java/br/edu/atitus/teamservice/controllers/TeamController.java
@@ -3,6 +3,7 @@ package br.edu.atitus.teamservice.controllers;
 import br.edu.atitus.teamservice.DTOs.TeamDTO;
 import br.edu.atitus.teamservice.entities.TeamEntity;
 import br.edu.atitus.teamservice.services.TeamService;
+import org.apache.tomcat.websocket.AuthenticationException;
 import org.flywaydb.core.internal.publishing.PublishingConfigurationExtension;
 import org.springframework.beans.BeanUtils;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +13,7 @@ import java.util.List;
 import java.util.UUID;
 
 @RestController
-@RequestMapping("/teams")
+@RequestMapping("/ws/teams")
 public class TeamController {
 
     private final TeamService teamService;
@@ -21,14 +22,8 @@ public class TeamController {
         this.teamService = teamService;
     }
 
-    @PostMapping("create")
-    public ResponseEntity<TeamEntity> createTeam(@RequestBody TeamDTO teamDTO) throws Exception {
-        TeamEntity newTeam = new TeamEntity();
-        BeanUtils.copyProperties(teamDTO, newTeam);
-
-        var createdTeam = teamService.createTeam(newTeam);
-
-        return ResponseEntity.ok(createdTeam);
+    private boolean validadeUserType(Long userType) {
+        return userType != 0;
     }
 
     @GetMapping("/{teamid}")
@@ -46,13 +41,42 @@ public class TeamController {
         return ResponseEntity.ok(teamService.getTeams());
     }
 
+    @PostMapping("create")
+    public ResponseEntity<TeamEntity> createTeam(
+            @RequestBody TeamDTO teamDTO,
+            @RequestHeader("X-User-Id") Long userId,
+            @RequestHeader("X-User-Email") String userEmail,
+            @RequestHeader("X-User-Type") Long userType
+    ) throws Exception {
+        if (validadeUserType(userType)) throw new AuthenticationException("Usuário não tem nível de acesso");
+
+        TeamEntity newTeam = new TeamEntity();
+        BeanUtils.copyProperties(teamDTO, newTeam);
+
+        var createdTeam = teamService.createTeam(newTeam);
+        return ResponseEntity.ok(createdTeam);
+    }
+
     @DeleteMapping("/delete/{teamId}")
-    public void deleteTeam(@PathVariable UUID teamId) throws Exception {
+    public void deleteTeam(
+            @PathVariable UUID teamId,
+            @RequestHeader("X-User-Id") Long userId,
+            @RequestHeader("X-User-Email") String userEmail,
+            @RequestHeader("X-User-Type") Long userType
+    ) throws Exception {
+        if (validadeUserType(userType)) throw new AuthenticationException("Usuário não tem nível de acesso");
         teamService.deleteTeamById(teamId);
     }
 
     @PutMapping("/update/{teamId}")
-    public TeamEntity updateTeam(@PathVariable UUID teamId, @RequestBody TeamDTO teamDTO) throws Exception {
+    public TeamEntity updateTeam(
+            @PathVariable UUID teamId,
+            @RequestBody TeamDTO teamDTO,
+            @RequestHeader("X-User-Id") Long userId,
+            @RequestHeader("X-User-Email") String userEmail,
+            @RequestHeader("X-User-Type") Long userType
+    ) throws Exception {
+        if (validadeUserType(userType)) throw new AuthenticationException("Usuário não tem nível de acesso");
         return  teamService.updateTeam(teamId, teamDTO);
     }
